hugging_Face_Deployment/Plant_Leaf_Disease_Detection_LightWeight_Model/requirements.txt
torch
torchvision
gradio
pandas
timm
Pillow
requests
